+++ app/src/main/java/com/dynamo/spacex/data/network/model/pastlaunch/Payload.kt
@@ -8,8 +8,8 @@ data class Payload(
     val orbit: String, // LEO
     val orbit_params: OrbitParams,
     val payload_id: String, // FalconSAT-2
-    val payload_mass_kg: Int, // 20
-    val payload_mass_lbs: Int, // 43
+    val payload_mass_kg: Double, // 20
+    val payload_mass_lbs: Double, // 43
     val payload_type: String, // Satellite
     val reused: Boolean // false
 )
